backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
pymupdf
chromadb
python-dotenv
google-generativeai
